feat: melhorias na listagem, criação do css global, entre outras melhorias e funcionalidades

+++ src/app/modules/author/author-list/author-list.ts
@@ -3,6 +3,7 @@ import { AuthorModel } from '../../../models/author.model';
 import { AuthorService } from '../../../services/author/author.service';
 import { RouterModule } from '@angular/router';
 import { CommonModule } from '@angular/common';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-author-list',
@@ -24,7 +25,27 @@ export class AuthorList implements OnInit {
   }
 
   public async deleteAuthor(id: number) {
-    await this.authorService.delete(id);
+    const result = await Swal.fire({
+      title: 'Tem certeza?',
+      text: 'Você não poderá reverter essa ação!',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#4caf50',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Sim, excluir!',
+      cancelButtonText: 'Cancelar'
+    });
+
+    if (result.isConfirmed) {
+      await this.authorService.delete(id);
+      await this.loadAuthors(); 
+      Swal.fire({
+        title: 'Excluído!',
+        text: 'O autor foi removido com sucesso.',
+        icon: 'success',
+        confirmButtonColor: '#4caf50'
+      });
+    }
   }
 
 }
